# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Beratungssuche REST-API",
    "description": "Backend der Beratungssuche TU Berlin.\n",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "matildafreiheit@tu-berlin.de"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/bachelorarbeit-c90/Bachelorarbeit/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    },
    {
      "url": "https://vm021.qu.tu-berlin.de:3000",
      "description": "VM TU Berlin"
    },
    {
      "url": "https://vm021.qu.tu-berlin.de:3000"
    }
  ],
...
  "paths": {
...
    "/angebot_tags/{angebotId}": {
      "get": {
        "tags": [
          "nm"
        ],
        "summary": "Tags für ein Angebot abrufen",
        "description": "Gibt die Tags für ein bestimmtes Angebot zurück oder alle, wenn keine `angebotId` angegeben wird.",
        "operationId": "getAngebotTags",
        "parameters": [
          {
            "name": "angebotId",
            "in": "path",
            "description": "ID des Angebots, für das die Tags abgerufen werden sollen (optional).",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tags erfolgreich abgerufen",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_16"
                },
                "example": {
                  "data": [
                    {
                      "AngebotID": 1,
                      "TagID": 5
                    },
                    {
                      "AngebotID": 1,
                      "TagID": 8
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Fehler beim Abrufen der Tags."
          }
        }
      }
    },
    "/angebot_zielgruppe/{angebotId}": {
      "get": {
        "tags": [
          "nm"
        ],
        "summary": "Zielgruppen für ein Angebot abrufen",
        "description": "Gibt die Zielgruppen für ein bestimmtes Angebot zurück oder alle, wenn keine `angebotId` angegeben wird.",
        "operationId": "getAngebotZielgruppen",
        "parameters": [
          {
            "name": "angebotId",
            "in": "path",
            "description": "ID des Angebots, für das die Zielgruppen abgerufen werden sollen (optional).",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Zielgruppen erfolgreich abgerufen",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_17"
                },
                "example": {
                  "data": [
                    {
                      "AngebotID": 1,
                      "ZielgruppeID": 3
                    },
                    {
                      "AngebotID": 1,
                      "ZielgruppeID": 7
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Fehler beim Abrufen der Zielgruppen."
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "inline_response_200_16": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/inline_response_200_16_data"
            }
          }
        }
      },
      "inline_response_200_17": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/inline_response_200_17_data"
            }
          }
        }
      },
...
      "inline_response_200_16_data": {
        "type": "object",
        "properties": {
          "AngebotID": {
            "type": "integer"
          },
          "TagID": {
            "type": "integer"
          }
        }
      },
      "inline_response_200_17_data": {
        "type": "object",
        "properties": {
          "AngebotID": {
            "type": "integer"
          },
          "ZielgruppeID": {
            "type": "integer"
          }
        }
      }
    },
...
  }
}